Moved scoring out to util to be reused
