+++ src/utils/index.js
@@ -3,5 +3,6 @@ export * from './Global'
 export * from './hooks/useLocallyPersistedReducer'
 export * from './hooks/useRouter'
 export * from './PrivateRoute'
+export * from './scoring'
 export * from './themes'
 export * from './typography'
